Added confirmation check to outbox mail deletion.

+++ Internship-7-EF-Dmail.Presentation/Actions/AuthenticatedUserActions/OutboxActions/HideMailAction.cs
@@ -25,6 +25,15 @@ namespace Internship_7_EF_Dmail.Presentation.Actions.AuthenticatedUserActions.Ou
             Console.Clear();
             WriteLine(Name);
 
+            if (_selected.Format == Data.Enums.MailFormat.Event)
+                WriteLine("Deleting this event from the outbox will not remove the event invitations you sent!", Style.Warning);
+            if (!GetConfirmation("Are you sure you want to delete this mail?", false))
+            {
+                WriteLine(OTHER_CANCELLED, Style.Emphasis);
+                WaitForInput();
+                return;
+            }
+
             Response response = _mailRepository.RemoveFromOutbox(_selected.Id);
 
             WriteGenericResponse(response);
